Use preprocessor directives for perf reasons, instead of an interface.
