+++ frameworks/CSharp/aspnetcore/PlatformBenchmarks/BenchmarkApplication.Caching.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+#if DATABASE
+
 using System.IO.Pipelines;
 using System.Threading.Tasks;
 
@@ -14,3 +16,5 @@ namespace PlatformBenchmarks
         }
     }
 }
+
+#endif
