+++ frameworks/CSharp/aspnetcore/PlatformBenchmarks/BenchmarkApplication.Fortunes.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+#if DATABASE
+
 using System.Collections.Generic;
 using System.IO.Pipelines;
 using System.Text.Encodings.Web;
@@ -51,3 +53,5 @@ namespace PlatformBenchmarks
         }
     }
 }
+
+#endif
