+++ frameworks/CSharp/aspnetcore/PlatformBenchmarks/BenchmarkApplication.MultipleQueries.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+#if DATABASE
+
 using System.IO.Pipelines;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -39,3 +41,5 @@ namespace PlatformBenchmarks
         }
     }
 }
+
+#endif
