+++ frameworks/CSharp/aspnetcore/PlatformBenchmarks/BenchmarkApplication.cs
@@ -44,7 +44,9 @@ namespace PlatformBenchmarks
         private readonly static AsciiString _fortunesTableEnd = "</table></body></html>";
         private readonly static AsciiString _contentLengthGap = new string(' ', 4);
 
-        public static IDb Db { get; set; }
+#if DATABASE
+        public static RawDb Db { get; set; }
+#endif
 
         [ThreadStatic]
         private static Utf8JsonWriter t_writer;
